+++ ready/AzNamingTool/Pages/AdminLog.razor
@@ -179,7 +179,7 @@
             await AdminLogService.DeleteAllItems();
             toastService.ShowSuccess("Admin Log purged!", "SUCCESS");
             AdminLogService.PostItem(new AdminLogMessage() { Title = "INFORMATION", Message = "Admin Log log purged." });
-            GeneralHelper.InvalidateCacheObject("AdminLogMessage");
+            CacheHelper.InvalidateCacheObject("AdminLogMessage");
             serviceResponse = await AdminLogService.GetItems();
             lstAdminLogMessages = (List<AdminLogMessage>)serviceResponse.ResponseObject;
             dataLoaded = true;
