+++ ready/AzNamingTool/Shared/EditModal.razor
@@ -126,7 +126,7 @@
             </div>
             <div class="modal-footer">
                 <button title="Save" @onclick="Save" class="btn btn-success">Update</button>
-                <button title="Cancel" @onclick="ModalInstance.CancelAsync" class="btn btn-secondary">Cancel</button>
+                <button title="Cancel" @onclick="Cancel" class="btn btn-secondary">Cancel</button>
             </div>
         </div>
     </div>
@@ -263,7 +263,7 @@
             {
                 if ((itemName != "") && (itemShortName != ""))
                 {
-                    if (GeneralHelper.ValidateShortName(itemShortName, type))
+                    if (ValidationHelper.ValidateShortName(itemShortName, type))
                     {
                         valid = true;
                         switch (type)
