+++ ready/AzNamingTool/Shared/MainLayout.razor
@@ -47,7 +47,7 @@
 <button onclick="topFunction()" id="btnScrollToTop" class="btn btn-success btn-sm" title="Go to top">Top</button>
 @if (PasswordModalOpen)
 {
-    <PasswordModal OnClose="@OnPasswordModalClose">
+    <PasswordModal OnClose="@OnPasswordModalClose" DisableBackgroundCancel="true">
     </PasswordModal>
 }
