+++ ready/AzNamingTool/Shared/PasswordModal.razor
@@ -56,7 +56,7 @@
         if (password != "")
         {
             // Make sure password is complex enough
-            if (GeneralHelper.ValidatePassword(password))
+            if (ValidationHelper.ValidatePassword(password))
             {
                 var config = GeneralHelper.GetConfigurationData();
 
@@ -96,7 +96,7 @@
         password = value;
 
         // Validate the password complexity
-        if (GeneralHelper.ValidatePassword(password))
+        if (ValidationHelper.ValidatePassword(password))
         {
             disabled = false;
         }
